apod - fix broken image metadata and error response / logging on room avatar event errors
There was a parameter off by one error. Propably introduced with the
room_send wrapper.

While testing the avatar feature I noticed that nothing happens if the bot is not at least a moderator in
the room. So I decided to send that info to the room and log the error
message as a warning.

The explanation was not altered by the apod module.
Turns out it was broken that day at the api level. See https://api.nasa.gov/planetary/apod?api_key=DEMO_KEY&hd=true&date=2022-09-13

fixes #216